Convert to go module
Removed glide files in favor of the go module descriptor files and adjusted the Makefile appropriately.

A copy of etcd client had to be kept in order to apply the patches required to the import paths.
Other minor changes had to be applied in the Go source files to make the tests pass.

The test script has been updated not to rely on hardcoded module name, but in the one set in go.mod.
Also, the go package names are obtained from 'go list' tool.