Add waitForRebootWindow to sleep till next window
waitForRebootWindow get the environment variable REBOOT_WINDOW_START and
REBOOT_WINDOW_LENGTH and check if everything is ok or not. If not then
return error. If ok then get the next periodic by the ParsePeriodic
function.

Then check if we are inside the reboot window or not. If we are inside
the reboot window then it will sleep until the next window and return a
true. If outside the window then it just return the false.

In the lockAndReboot function we check if it has sleep or not and handle the
error.