locksmithcl/locksmithcl: fix `defaultEnpoints`
Using etcd/v3, it seems we don't need to specify the scheme
(`http`/`https`).

If it's specified, it will use the ones provided by default and stick to
it - even if it's HTTP and we have a HTTPS configured etcd.

So it will fail because we send http request to a https service

Signed-off-by: Mathieu Tortuyaux <mathieu@kinvolk.io>