locksmithcl: reset endpoints when we set a new one
This behavior was working fine previously but no more with the etcd/v2
upgrade.

In the case we have an endpoint like `https://127.0.0.1:2379`, this one
will be _added_ to the list of `defaultEndpoints` which lead to this
endpoints:
```go
[]string{
	"http://127.0.0.1:2379",
	"http://127.0.0.1:4001",
	"https://127.0.0.1:2379",
}
```

It's not an issue when we have a retry mechanism - but in this case if
`etcd` use HTTPS and we try an `HTTP` endpoint, we will get an io.EOF
error.

Signed-off-by: Mathieu Tortuyaux <mathieu@kinvolk.io>