systemd: Restore access to TTYs for reboot delay
When sessions are active, locksmith write a message to the TTYs of the
sessions and then delays the reboot for 5 minutes to give the user time
to stop the reboot or finish the work.
The commit for cgroup memory and CPU limits also brought in a change to
disallow /dev/tty* access which broke the delay for console users except
SSH sessions which are under /dev/pts/*.
Allow device access to have a delay of 5 minutes when sessions are
active. This includes the autologin session even if no interaction was
done there recently. Still, the reboot delay doesn't hurt and since
update-engine has a random delay for pulling updates, there is no big
difference in the end.